refactor: rename SocialMediaButton-styling file
